refactor(gateway): Q11.g — extract rest_v1/tasks/attachments.rs (GAR-658)
Extracts task attachment section from monolithic rest_v1/tasks/mod.rs into
rest_v1/tasks/attachments.rs (~383 LOC). Reduces mod.rs from ~1537 lines
to ~1153 lines; re-exports via pub use keep all router call-sites unchanged.
No openapi.rs change needed — attachment handlers not registered in paths!().
Closes GAR-658 (slice 7 of Q11 / GAR-635).
